Craiglist-Scrapper/requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
pandas==2.1.3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
twilio==8.10.0
flask==3.0.0
openai==1.3.0
python-dotenv==1.0.0
asgiref==3.7.0
uvicorn==0.24.0
